Rename invert.py to _invert.py
