+++ skimage/util/_invert.py
